% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_climate_on_firm_regs.R
\name{climate_on_firm_regs_charts}
\alias{climate_on_firm_regs_charts}
\title{Charts showing impact of climate variables on firm outcomes}
\usage{
climate_on_firm_regs_charts(data, level_or_difference, output_directory)
}
\arguments{
\item{data}{Regression-ready dataset prepared by 1_prep_reg_data.}

\item{level_or_difference}{Either "Level" or "Difference"; indicates whether for the given climate variable, its survey year level or the difference of this level from its long-run mean should be used.}

\item{output_directory}{Folder path for where charts will be saved.}
}
\value{
One chart for heat days, one for temperature, and one for temperature volatility saved in the output_directory.
}
\description{
The climate_on_firm_regs_charts() function exports charts showing results from the regressions of firm outcomes on climate variables.
}
\note{
Remember to use forward slashes "/" in the folder path provided.
}
