% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_subset_data.R
\name{data_subset}
\alias{data_subset}
\title{Subset master geo-linked WBES dataset based on survey, country, or region.}
\usage{
data_subset(data_file_path, filter_value, subset_type)
}
\arguments{
\item{data_file_path}{A string of the user's file path to the raw data.}

\item{filter_value}{A string that indicates the value of the country, survey (CountryYear), or region the user wants to analyze.}

\item{subset_type}{A string that indicates the type of subsetting: "survey", "country", or "region".}
}
\value{
The subsetted dataset as a dataframe (save to object and use as input in other wbfirmadaptation functions).
}
\description{
The data_subset() function subsets the raw master dataset based on the user's chosen type: 'survey', 'country', or 'region'.
}
