% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_prep_reg_data.R
\name{prep_reg_data}
\alias{prep_reg_data}
\title{Process subset of the master geo-linked WBES dataset to make it regression-ready.}
\usage{
prep_reg_data(subset_data)
}
\arguments{
\item{subset_data}{An R dataframe containing the subset of the raw master dataset to be analyzed.}
}
\value{
The regression-ready dataset as a dataframe (save to object and use as input in other wbfirmadaptation functions).
}
\description{
The prep_reg_data() function processes the subset of the raw master dataset (produced by data_subset()) and produces a regression-ready dataset.
}
