% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_climate_on_firm_regs.R
\name{climate_on_firm_regs_tables}
\alias{climate_on_firm_regs_tables}
\title{Tables showing impact of climate variables on firm outcomes}
\usage{
climate_on_firm_regs_tables(data, level_or_deviation, output_directory)
}
\arguments{
\item{data}{Regression-ready dataset prepared by 1_prep_reg_data.}

\item{level_or_deviation}{Either "level" or "deviation"; indicates whether for the given climate variable, its survey year level or the deviation of this level from its long-run mean should be used.}

\item{output_directory}{Folder path for where tables will be saved.}
}
\value{
One table for heat days, one for temperature, and one for temperature volatility saved in the output_directory.
}
\description{
The climate_on_firm_regs_tables() function exports tables showing results from the regressions of firm outcomes on climate variables.
}
\note{
Remember to use forward slashes "/" in the folder path provided.
}
