% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_interaction_regs.R
\name{interaction_reg_table}
\alias{interaction_reg_table}
\title{Table showing impact of the interaction between firm characteristics and climate variables on firm outcomes.}
\usage{
interaction_reg_table(
  reg_data,
  firm_outcome,
  climate_var,
  level_or_difference,
  firm_characteristics,
  html_or_tex,
  filename,
  output_directory
)
}
\arguments{
\item{firm_outcome}{Dependent variable passed as a string, see guidance note for options.}

\item{climate_var}{Independent climate variable passed as a string, options include "Temperature", "TemperatureVolatility", and "HeatDays".}

\item{level_or_difference}{Either "Level" or "Difference"; indicates whether for the given climate variable, its survey year level or the difference of this level from its long-run mean should be used.}

\item{firm_characteristics}{Independent firm characteristic/policy variable passed as a string or list of strings, see guidance note for options.}

\item{html_or_tex}{Either "html" or "tex"; indicates whether the table will be saved in .html or .tex format.}

\item{filename}{Name of the file to be saved, excluding the file extension.}

\item{output_directory}{File path for where the table will be saved, ending in .html e.g. "C:/Users/Nolan/Project/table.html".}

\item{data}{Regression-ready dataset prepared by 1_prep_reg_data.}
}
\value{
Regression results table for the variables provided saved in the output_directory.
}
\description{
Exports a regression result table featuring interactions between the firm characteristics and climate variables provided.
}
\note{
Remember to use forward slashes "/" in the folder path.
}
