% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_prep_reg_data.R
\name{prep_reg_data}
\alias{prep_reg_data}
\title{Process master geo-linked WBES dataset to make it regression-ready.}
\usage{
prep_reg_data(data_directory, survey)
}
\arguments{
\item{data_directory}{A string of the user's directory containing the raw data.}

\item{survey}{A string structured CountryYear indicating the survey the user wants to analyze.}
}
\value{
The regression-ready dataset as a dataframe (save to object and use as input in other wbfirmadaptation functions).
}
\description{
The prep_reg_data() function processes the raw master dataset and produces a regression-ready dataset.
}
\note{
"master.dta" (WBES geo-linked master dataset) must be saved in the project directory.
}
